refactor: split materialize modules
