+++ packages/server/src/entities/materialize-json.ts
@@ -1,0 +1,35 @@
+export function readJsonObject(raw: string | null): Record<string, unknown> {
+  if (!raw) return {};
+  try {
+    const parsed = JSON.parse(raw);
+    return parsed && typeof parsed === "object" ? (parsed as Record<string, unknown>) : {};
+  } catch {
+    return {};
+  }
+}
+
+export function readPersonEmailFromMetadata(metadata: string | null): string | null {
+  if (!metadata) return null;
+  try {
+    const parsed = JSON.parse(metadata) as { email?: unknown };
+    if (typeof parsed.email === "string" && parsed.email.length > 0) return parsed.email.toLowerCase();
+  } catch {
+    return null;
+  }
+  return null;
+}
+
+export function parseAliasesString(aliases: string | null): string[] {
+  if (!aliases) return [];
+  try {
+    const parsed = JSON.parse(aliases);
+    if (Array.isArray(parsed)) return parsed.filter((v): v is string => typeof v === "string");
+  } catch {
+    return [];
+  }
+  return [];
+}
+
+export function isString(value: unknown): value is string {
+  return typeof value === "string";
+}
